requirements.txt
absl-py==0.12.0
astunparse==1.6.3
bleach==1.5.0
blurhash==1.1.4
boost==0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
DBUtils==2.0.1
decorator==4.4.2
distro==1.5.0
Django==2.0
django-cors-headers==2.0.0
dlib==19.22.0
enum34==1.1.10
face-alignment==1.0.0
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
futures==3.1.1
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
Mastodon.py==1.5.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==3.4.6.27
opt-einsum==3.3.0
Pillow==8.2.0
pip==21.1.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.22
pytz==2021.1
PyWavelets==1.1.1
PyYAML==4.2b1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scipy==1.5.4
setuptools==56.0.0
six==1.15.0
SQLAlchemy==1.4.12
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tf-slim==1.1.0
tifffile==2020.9.3
torch==1.1.0
torchvision==0.3.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
zipp==3.4.1
